backend/python-services/ai-service/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# AI/ML libraries
openai==1.3.8
langchain==0.0.348
langchain-openai==0.0.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# Database connectivity
sqlite3
asyncpg==0.29.0
redis==5.0.1

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.1.0
jinja2==3.1.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Food/recipe specific libraries
spacy==3.7.2
recipe-scrapers==14.48.0
nltk==3.8.1